requirements.txt
python-dotenv==0.19.2
tweepy==4.10.0
solana==0.30.2
pandas==1.4.2
numpy==1.22.3
scikit-learn==1.0.2
requests==2.27.1
beautifulsoup4==4.11.1
pymongo==4.1.1
python-telegram-bot==13.11
anchorpy==0.14.0
solders==0.18.1